Art and Movement
Added the art for the player to the project and attached it to the player object(which was also created).

Added functionality so the player can move left and right.

May add feature where the player changes when moving left and right. Also, might work on how the speed works.